+++ backend/test/auth.test.js
@@ -1,0 +1,17 @@
+const chai = require('chai');
+const chaiHttp = require('chai-http');
+const mongoose = require('mongoose');
+const User = require('../models/User');
+const sinon = require('sinon');
+const { register , getUserHelpPosts, updateHelpPost, deleteHelpPost } = require('../controllers/authController');
+const UserRepository = require('../repositories/UserRepository');
+
+const { expect } = chai;
+
+
+
+describe('Basic Mocha Test', () => {
+  it('should pass a simple assertion', () => {
+    expect(1 + 1).to.equal(2);
+  });
+});
